% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCM_Ranking.R
\name{bcm_ranking}
\alias{bcm_ranking}
\title{Rank plot of batch-correction methods provided by SelectBCM tool}
\usage{
bcm_ranking(rank_data)
}
\arguments{
\item{diagnostic_matirx}{diagnostic_matirx obtained from the previous step `diagnostic_matrix.microarray`.}
}
\description{
Rank plot of batch-correction methods provided by SelectBCM tool
}
