% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComBat_seq.R
\name{ComBat_seq}
\alias{ComBat_seq}
\title{Adjust for batch effects using an empirical Bayes framework in RNA-seq raw counts}
\usage{
ComBat_seq(
  counts,
  batch,
  group = NULL,
  covar_mod = NULL,
  full_mod = TRUE,
  shrink = FALSE,
  shrink.disp = FALSE,
  gene.subset.n = NULL
)
}
\arguments{
\item{counts}{Raw count matrix from genomic studies (dimensions gene x sample)}

\item{batch}{Batch covariate (only one batch allowed)}

\item{group}{Vector / factor for condition of interest}

\item{covar_mod}{Model matrix for other covariates to include in linear model besides batch and condition of interest}

\item{full_mod}{Boolean, if TRUE include condition of interest in model}

\item{shrink}{Boolean, whether to apply empirical Bayes estimation on parameters}

\item{shrink.disp}{Boolean, whether to apply empirical Bayes estimation on dispersion}

\item{gene.subset.n}{Number of genes to use in empirical Bayes estimation, only useful when shrink = TRUE}
}
\value{
data A probe x sample count matrix, adjusted for batch effects.
}
\description{
ComBat_seq is an extension to the ComBat method using Negative Binomial model.
}
\examples{



}
