% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.evaluation.R
\name{detect_effect}
\alias{detect_effect}
\title{Detecting batch-effect in raw merged dataset}
\usage{
detect_effect(result, experiment, batch.factors, N1, N2, filter)
}
\arguments{
\item{result}{A merged experiment without batch correction obtained from step ('merge_experiments').}

\item{experiment}{A merged experiment without batch correction obtained from step ('merge_experiments').}

\item{batch.factors}{A list of factors to perform PVCA analysis. Along with the batch factor, one biological factor
which can be used to assess over-fitting should be provided.}

\item{N1}{is the number of randomly picked cells for the BatchEntropy function.}

\item{N2}{is the number of nearest nearest neighbors of the sample (from all batches) to check (for BatchEntropy function).}

\item{filter}{A string. Should be one of following string- 'symbol', 'ensembl_gene_id', or 'entrezgene_id'
depending on gene label for the given dataset.}
}
\value{
A list of the evaluation methods on the batch-corrected experiment.
}
\description{
This is an accessory function that performs a subset of evaluation tests of
 'evaluation_matrix' function and provides estimates whether the merged dataset obtained
 after 'merge_experiments' requires batch correction or not. A higher value of pvca.batch, silhouette,
 pcRegression, and entropy is indicative of batch-effects in a raw merged dataset without
 having any correction.
}
\examples{
 detect_effect(experiments,experiment = experiments, batch.factors=c("batch","Disease"),10,10,'symbol')
}
