% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostic.microarray.R
\name{diagnostic_matrix.microarray}
\alias{diagnostic_matrix.microarray}
\title{Diagnostic plot and ranking of batch-correction methods}
\usage{
diagnostic_matrix.microarray(evaluation)
}
\arguments{
\item{evaluation}{evaluation is the evaluation list obtained from the previous step `evaluation_matrix`.}
}
\value{
A list of two data-frame- (a) raw - simple data-frame output of evaluation matrix and, (b) ranked-
Ranked data-frame of evaluation matrix which has additional column `sumRank` containing final Rank of each method.
Ranks are in descending performance order, i.e. method having score 1 will be the best method.
}
\description{
The function performs ranking and plotting of the evaluation data obtained at
the previous step 'batch_evaluation’. Here, methods are ranked on their individual performance,
and finally 'sumRank' is the final Rank of each method for the given dataset.
Rank 1 will be the best performer method.
}
