% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostic_matrix.RNAseq.R
\name{diagnostic_matrix.RNAseq}
\alias{diagnostic_matrix.RNAseq}
\title{Ranking of batch-correction methods for bulk RNAseq experiments}
\usage{
diagnostic_matrix.RNAseq(evaluation)
}
\arguments{
\item{evaluation}{evaluation is the evaluation list obtained from the previous step `evaluation_matrix`.}
}
\value{
A list of two data-frame- (a) raw - simple data-frame output of evaluation matrix and, (b) ranked-
Ranked data-frame of evaluation matrix which has additional column `sumRank` containing final Rank of each method.
Ranks are in descending performance order, i.e. method having score 1 will be the best method.
}
\description{
The function performs ranking of the evaluation data obtained at
the previous step 'batch_evaluation’ for the bulk RNAseq experiments. Here, methods are ranked on their individual performance,
and finally 'sumRank' is the final Rank of each method for the given dataset.
Rank 1 will be the best performer method.
}
\examples{
 diagnostic.RNAseq(assessment)
}
