% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plot.RNAseq.R
\name{diagnostic_plot.RNAseq}
\alias{diagnostic_plot.RNAseq}
\title{plotting of ranks obatined form diagnostic_matrix.RNAseq step}
\usage{
diagnostic_plot.RNAseq(diagnostic_matirx)
}
\arguments{
\item{diagnostic_matirx}{diagnostic_matirx obtained from the previous step `diagnostic_matrix.RNAseq`.}
}
\value{
A list of two plots-
 (a) Output is `diagnostic_plot`, where the x-axis is the evaluation methods and the y-axis is the Rank of each normalization method.
  (b) Additional plot to summarise rankings for each method as a violin plot including summary statistics for each BCM.
}
\description{
The function performs plotting of the evaluation data obtained at
the previous step 'diagnostic_matrix.RNAseq’.
}
\examples{
 diagnostic_plot.RNAseq(diagnostic_matirx)
}
