% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadatacreation.R
\name{download_experiments_from_ExpressionAtlas}
\alias{download_experiments_from_ExpressionAtlas}
\title{download_experiments_from_ExpressionAtlas}
\usage{
download_experiments_from_ExpressionAtlas(
  ...,
  destdir = getwd() \%>\% paste("experiments", sep = "/")
)
}
\arguments{
\item{experiments}{Character list of Expression atlas experiment id}
}
\value{
A list of the downloaded experiments.
}
\description{
To use data from Expression Atlas that can be downloaded in `.Rdata` format.
}
\details{
The experiment should be either bulk-RNAseq(SummarizedExperiment) or microarray(ExpressionSet).
WARNING: Different types of experiment should not be processed together. For example, one SummarizedExperiment with expressionSet experiment, otherwise it will cause an error in further steps.
}
\examples{
experiments <- download_experiments_from_ExpressionAtlas('E-ERAD-169','E-GEOD-73175','E-MTAB-2801')
}
