% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadatacreation.R
\name{load_experiments}
\alias{load_experiments}
\title{Loading of experiments in a list}
\usage{
load_experiments(directory, names = dir(directory), item.SimpleList = "rnaseq")
}
\arguments{
\item{directory}{A character string that is the path to the directory where all the experiments to load, and not any other file, are stored in Rdata format.}

\item{item.SimpleList}{If the experiments are wrapped in a SimpleList object, this is the accessor (character string or integer) of the actual experiment object within this SimpleList. The experiment is unwrapped from this SimpleList.}
}
\value{
A list of the loaded experiments.
}
\description{
Loading data in R- If the data files are already on your computer, you can use this step.
}
\details{
The experiment should be either bulk-RNAseq(Summarized experiment) or microarray(ExpressionSet).
WARNING: Different types of experiment should not be processed together. For example, one Summarized experiment with expressionSet experiment, otherwise it will cause an error in further steps.
}
