% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadatacreation.R
\name{merge_experiments}
\alias{merge_experiments}
\title{Rough merger of experiments}
\usage{
merge_experiments(
  experiments,
  filter.unexpressed.genes = TRUE,
  log,
  force = FALSE
)
}
\arguments{
\item{experiments}{A list of wrapped experiments.}

\item{filter.unexpressed.genes}{A logical indicating if the genes that are unexpressed across all the samples of a batch shall be removed. This is for the proper functioning of the current version of ComBat in correct_batch_effect.}

\item{log}{A logical ndicating of whether the data shall be log-transformed(recommended).}

\item{force}{A logical. If the experiments of the list are of different classes, the function is stopped, unless force=TRUE.}
}
\value{
Single SummarizedExperiment or ExpressionSet
}
\description{
The function merge_experiments merges all the experiments in the list in a single experiment object and doesn't perform any correction. This function has two additional arguments log and filter (set to TRUE by default).
It is recommended to give microarray data only with back-ground correction and after filtering low-expressed genes
}
\note{
Caution: Sometime during merging experiments, phenodata (SDRF) file gets corrupted, it is advised to always check meta-data before proceeding further.
}
