% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadatacreation.R
\name{remove_isolated_experiments}
\alias{remove_isolated_experiments}
\title{Removal of isolated experiments}
\usage{
remove_isolated_experiments(experiments, biological.group)
}
\arguments{
\item{experiments}{A list of wrapped experiments.}

\item{biological.group}{A character string indicating the biological covariate that makes the biological groups. This must be the name of a column of the experiment.}
}
\value{
A list of wrapped experiments.
}
\description{
To correct the batch effect, one needs to take the biological characteristics of the samples into account. If no sample of an experiment shares biological characteristics with samples from other batches, it is not possible to correct the batch effect with these batches since one cannot distinguish the biological difference from the artifact. The function remove_isolated_experiments removes the isolated experiments and plots graphs of intersections between the experiments before and after removal.
}
\details{
The experiments of the list that have no biological group in common with the other experiments are removed. This is an essential step to the integration of experiments using batch effect correction. Without it, the biological groups that are unshared between the experiments are confounding factors with the batch effect, and the latter cannot be corrected. The ultimate condition for this is the graph of biological intersections between the experiments to be a connected graph.

WARNING: this function only checks and removes the experiments that are isolated. However, there can remain several clusters of experiments that are part of a connected subgraph, while the graph of all the experiments isn't connected. The graph of intersections after the removal of experiments is displayed so that the user can notice it. In this condition, the user has to choose manually some experiments that make a connected subgraph.
}
